# ODataUriSlim struct