--- a/src/Microsoft.OData.Core/Evaluation/ODataMetadataContext.cs
+++ b/src/Microsoft.OData.Core/Evaluation/ODataMetadataContext.cs
@@ -36,7 +36,7 @@ namespace Microsoft.OData.Evaluation
         /// <summary>
         /// Gets the OData uri.
         /// </summary>
-        ODataUriSlim ODataUri { get; }
+        ODataUriSlim? ODataUri { get; }
 
         /// <summary>
         /// Gets an entity metadata builder for the given resource.
@@ -95,7 +95,7 @@ namespace Microsoft.OData.Evaluation
         /// <summary>
         /// The OData Uri.
         /// </summary>
-        private readonly ODataUriSlim odataUri;
+        private readonly ODataUriSlim? odataUri;
 
         /// <summary>
         /// The service base Uri.
@@ -115,7 +115,7 @@ namespace Microsoft.OData.Evaluation
         /// <param name="metadataDocumentUri">The metadata document uri.</param>
         /// <param name="odataUri">The request Uri.</param>
         /// <remarks>This overload should only be used by the writer.</remarks>
-        public ODataMetadataContext(bool isResponse, IEdmModel model, Uri metadataDocumentUri, in ODataUriSlim odataUri)
+        public ODataMetadataContext(bool isResponse, IEdmModel model, Uri metadataDocumentUri, in ODataUriSlim? odataUri)
             : this(isResponse, /*OperationsBoundToEntityTypeMustBeContainerQualified*/ null, EdmTypeWriterResolver.Instance, model, metadataDocumentUri, in odataUri)
         {
         }
@@ -136,7 +136,7 @@ namespace Microsoft.OData.Evaluation
             EdmTypeResolver edmTypeResolver,
             IEdmModel model,
             Uri metadataDocumentUri,
-            in ODataUriSlim odataUri)
+            in ODataUriSlim? odataUri)
         {
             Debug.Assert(edmTypeResolver != null, "edmTypeResolver != null");
             Debug.Assert(model != null, "model != null");
@@ -166,7 +166,7 @@ namespace Microsoft.OData.Evaluation
             EdmTypeResolver edmTypeResolver,
             IEdmModel model,
             Uri metadataDocumentUri,
-            in ODataUriSlim odataUri,
+            in ODataUriSlim? odataUri,
             JsonLightMetadataLevel metadataLevel)
             : this(isResponse, operationsBoundToEntityTypeMustBeContainerQualified, edmTypeResolver, model, metadataDocumentUri, in odataUri)
         {
@@ -215,7 +215,7 @@ namespace Microsoft.OData.Evaluation
         /// <summary>
         /// Gets the OData uri.
         /// </summary>
-        public ODataUriSlim ODataUri
+        public ODataUriSlim? ODataUri
         {
             get
             {