--- a/src/Microsoft.OData.Core/JsonLight/JsonNoMetadataLevel.cs
+++ b/src/Microsoft.OData.Core/JsonLight/JsonNoMetadataLevel.cs
@@ -79,7 +79,7 @@ namespace Microsoft.OData.JsonLight
             SelectedPropertiesNode selectedProperties,
             bool isResponse,
             bool keyAsSegment,
-            ODataUri odataUri,
+            in ODataUriSlim odataUri,
             ODataMessageWriterSettings settings)
         {
             return ODataResourceMetadataBuilder.Null;