--- a/test/FunctionalTests/Microsoft.OData.Core.Tests/Evaluation/ODataMissingOperationGeneratorTests.cs
+++ b/test/FunctionalTests/Microsoft.OData.Core.Tests/Evaluation/ODataMissingOperationGeneratorTests.cs
@@ -211,7 +211,7 @@ namespace Microsoft.OData.Tests.Evaluation
             throw new NotImplementedException();
         }
 
-        public ODataUri ODataUri { get; set; }
+        public ODataUriSlim? ODataUri { get; set; }
     }
 
     public class AllPropertiesComparer<T> : IEqualityComparer<T>