… src/Microsoft.OData.Core/ODataWriterCore.cs
… test/FunctionalTests/Microsoft.OData.Core.Tests/ODataWriterCoreTests.cs